requirements.txt
playwright
beautifulsoup4
requests
google-generativeai
python-dotenv
